Harden A–E module migration compatibility and Complaint FK type safety

+++ Modules/Complaint/Database/Migrations/2026_04_10_600001_add_complaint_overlay_columns_to_tickets_table.php
@@ -42,7 +42,7 @@ return new class extends Migration
             }
 
             if (!Schema::hasColumn('tickets', 'assigned_cleaner_id')) {
-                $table->unsignedBigInteger('assigned_cleaner_id')->nullable()->after('refund_amount');
+                $table->unsignedInteger('assigned_cleaner_id')->nullable()->after('refund_amount');
             }
 
             if (!Schema::hasColumn('tickets', 'complaint_source')) {
@@ -59,7 +59,7 @@ return new class extends Migration
             }
 
             if (!Schema::hasColumn('tickets', 'resolved_by')) {
-                $table->unsignedBigInteger('resolved_by')->nullable()->after('resolved_at');
+                $table->unsignedInteger('resolved_by')->nullable()->after('resolved_at');
             }
         });
